requirements.txt
keras
pillow
google.colab